indexer/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[tool.poetry]
name = "indexer"
version = "0.0.0"
description = "Indexer for the documents"
authors = [ "Jacob Petterle <jacobpetterle@tai-tutor.team>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-lambda-powertools = "^2.43.1"
  boto3 = "^1.35.2"
  pinecone-client = "^5.0.1"
  pyarrow = "^17.0.0"
  pydantic-settings = "^2.4.0"
  pydantic = "^2.8.0"
  python = "^3.9"
  urllib3 = ">=1.26.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
requests = "^2.26.0"

  [tool.poetry.group.dev.dependencies.boto3-stubs]
  version = "^1.34.105"
  extras = [ "s3", "bedrock-runtime" ]

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
